requirements.txt
asgiref==3.11.1
Django==6.0.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
greenlet==3.3.2
pillow==12.2.0
playwright==1.58.0
pyee==13.0.1
sqlparse==0.5.5
typing_extensions==4.15.0
django-taggit==4.0.0
